Refactor trial_wf.
