+++ recirq/qcqmc/bitstrings.py
